# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minions Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, cheapest, expensive, popular]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [products]
      summary: Get user favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List user deals
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller, all]
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Deals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealsListResponse"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal (buy product)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
  /messages/chats:
    get:
      operationId: getChats
      tags: [messages]
      summary: Get chat list
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
  /messages/{userId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages with user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send message
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
...
  /admin/deals:
    get:
      operationId: adminListDeals
      tags: [admin]
      summary: List all deals (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealsListResponse"
...
  /admin/categories:
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        bio:
          type: string
        balance:
          type: number
        frozenBalance:
          type: number
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        totalVolume:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        sellerLevel:
          type: string
        telegramId:
          type: string
        refCode:
          type: string
        createdAt:
          type: integer
        lastActive:
          type: integer
      required: [id, username]
...
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        bio:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        productCount:
          type: integer
        sortOrder:
          type: integer
      required: [id, name, slug]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
      required: [name, slug]
    Product:
      type: object
      properties:
        id:
          type: string
        sellerId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        categoryName:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        game:
          type: string
        server:
          type: string
        status:
          type: string
        views:
          type: integer
        soldCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        isFavorited:
          type: boolean
        isPromoted:
          type: boolean
        createdAt:
          type: integer
        updatedAt:
          type: integer
      required: [id, sellerId, title, price, category]
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        deliveryData:
          type: string
        game:
          type: string
        server:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, description, price, category]
...
    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total]
    Deal:
      type: object
      properties:
        id:
          type: string
        dealNumber:
          type: integer
        buyerId:
          type: string
        buyer:
          $ref: "#/components/schemas/User"
        sellerId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        amount:
          type: number
        sellerAmount:
          type: number
        commission:
          type: number
        status:
          type: string
          enum: [pending, paid, delivered, completed, disputed, cancelled, refunded]
        deliveryData:
          type: string
        buyerConfirmed:
          type: boolean
        autoCompleteAt:
          type: integer
        disputeReason:
          type: string
        createdAt:
          type: integer
      required: [id, buyerId, sellerId, productId, amount, status]
    CreateDealBody:
      type: object
      properties:
        productId:
          type: string
      required: [productId]
...
    DealsListResponse:
      type: object
      properties:
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [deals, total]
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        receiverId:
          type: string
        text:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: integer
      required: [id, senderId, text]
    SendMessageBody:
      type: object
      properties:
        text:
          type: string
      required: [text]
    Chat:
      type: object
      properties:
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
      required: [userId]
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total]
...